requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
Django==2.0.9
django-extensions==2.2.9
django-simple-history==2.10.0
djangorestframework==3.11.0
drf-extensions==0.6.0
idna==2.6
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5.dev0
pylint==2.4.4
pytz==2019.3
requests==2.18.4
six==1.14.0
sqlparse==0.3.1
urllib3==1.22
wrapt==1.11.2
xmltodict==0.12.0
django-rest-swagger==2.2.0
django-allauth
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0